Make CTRL runtime portable to Linux 64-bit
Platform abstraction layer:
- Add Platform.h with IRAM_ATTR, getTime(), freeHeap(), cpuFreq(),
  rssi(), and log() abstractions
- Add ESP32Platform.cpp (wraps ESP-IDF/Arduino APIs)
- Add LinuxPlatform.cpp (uses std::chrono, stdio)

IOValue portability fix:
- Remove IOValue* ref pointer from the 4-byte IOValue union
- Add separate IOValue** inputRefs array to FunctionBlock
- IOValue stays exactly 4 bytes on all platforms
- All function library casts continue to work unchanged

Core cleanup:
- Replace ESP.getFreeHeap/getCpuFreqMHz/esp_timer_get_time/WiFi.RSSI
  with Platform:: calls in Controller
- Replace Serial.printf with Platform::log in Link
- Replace Arduino min/max macros with std::min/std::max
- Replace Arduino boolean with bool, uint with unsigned int
- Replace VLAs in Link.cpp with std::vector
- Fix || vs | bug in disableMonitoring()
- Clean up include dependencies (remove unnecessary Link.h includes)
- Use uintptr_t (ptr_t) for wire protocol pointer fields
- Add static to namespace-scoped names[] arrays in function libs

Build system:
- Add CMakeLists.txt for Linux build (C++14)
- Add test/test_main.cpp with 10 tests covering IOValue size,
  function blocks, connections, disconnect, type conversion,
  logic, circuits, cyclic tasks, timers, and platform time
- Add .gitignore for build/

+++ src/CTRL/CyclicTask.cpp
@@ -1,6 +1,7 @@
 #include "CyclicTask.h"
+#include "Controller.h"
+#include "FunctionBlock.h"
 #include "Circuit.h"
-#include "Esp.h"
 
 CyclicTask::CyclicTask(Controller* controller, uint32_t interval_ms, uint32_t offset_ms) :
     controller (controller),
@@ -80,7 +81,7 @@ void CyclicTask::setOffset(uint32_t time) {
 }
 
 void CyclicTask::addFunction(FunctionBlock* func, int32_t index) {
-    if (index > 0 && index < funcList.size()) {
+    if (index > 0 && index < (int32_t)funcList.size()) {
         funcList.insert(funcList.begin() + index, func);
     } else
         funcList.push_back(func);
